+++ app/src/main/java/com/example/movielister/login/LoginViewModel.kt
@@ -2,32 +2,23 @@ package com.example.movielister.login
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import com.example.movielister.model.SessionModel
-import com.example.movielister.model.TokenModel
-import com.example.movielister.network.LoginNetwork
-import kotlinx.coroutines.flow.MutableSharedFlow
+import com.example.movielister.repository.LoginRepository
 import kotlinx.coroutines.flow.asSharedFlow
 import kotlinx.coroutines.launch
 
 class LoginViewModel : ViewModel() {
-    private val loginService = LoginNetwork.createLoginAPI()
+    private val loginRepository = LoginRepository()
 
-    private val _token = MutableSharedFlow<TokenModel>()
-    val token = _token.asSharedFlow()
+    val token = loginRepository._token.asSharedFlow()
 
-    private val _sessionModel = MutableSharedFlow<SessionModel>()
-    val sessionModel = _sessionModel.asSharedFlow()
+    val sessionModel = loginRepository._sessionModel.asSharedFlow()
 
-    private val _loginToken = MutableSharedFlow<TokenModel>()
-    val loginToken = _loginToken.asSharedFlow()
+    val loginToken = loginRepository._loginToken.asSharedFlow()
 
     fun createToken() {
         viewModelScope.launch {
             try {
-                loginService.createToken()?.let {
-                    val token = it
-                    _token.emit(token)
-                }
+                loginRepository.createToken()
             } catch (e: Exception) {
                 e.printStackTrace()
             }
@@ -37,10 +28,7 @@ class LoginViewModel : ViewModel() {
     fun createSession(token: String) {
         viewModelScope.launch {
             try {
-                loginService.createSession(token)?.let {
-                    val sessionModel = it
-                    _sessionModel.emit(sessionModel)
-                }
+                loginRepository.createSession(token)
             } catch (e: Exception) {
                 e.printStackTrace()
             }
@@ -50,10 +38,7 @@ class LoginViewModel : ViewModel() {
     fun createSessionWithLogin(username: String, password: String, token: String) {
         viewModelScope.launch {
             try {
-                loginService.createSessionWithLogin(username, password, token)?.let {
-                    val loginToken = it
-                    _loginToken.emit(loginToken)
-                }
+                loginRepository.createSessionWithLogin(username, password, token)
             } catch (e: Exception) {
                 e.printStackTrace()
             }
