--- app/src/main/java/com/example/movielister/network/LoginNetwork.kt
@@ -1,9 +1,0 @@
-package com.example.movielister.network
-
-import com.example.movielister.api.LoginAPI
-
-object LoginNetwork {
-    fun createLoginAPI(): LoginAPI {
-        return WebService.createRetrofit().create(LoginAPI::class.java)
-    }
-}
